third_party/WebKit/LayoutTests/inspector/sources/debugger/debug-inlined-scripts-fragment-id.html

<html>
<head>
<script>
function f4()
{
    return 0;
}
f4();
</script>

<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>

<script>

var test = function()
{
    InspectorTest.startDebuggerTest(step0, true);

    function step0()
    {
        InspectorTest.evaluateInPage("window.location=\"#hash\"", step1);
    }

    function step1(loc)
    {
        InspectorTest.addResult("window.location: " + loc.description);
        InspectorTest.showScriptSource("debug-inlined-scripts-fragment-id.html", step2);
    }

    function step2(sourceFrame)
    {
        InspectorTest.addResult("Script source was shown.");
        InspectorTest.setBreakpoint(sourceFrame, 5, "", true);
        InspectorTest.waitUntilPaused(step3);
        InspectorTest.reloadPage(InspectorTest.completeDebuggerTest.bind(InspectorTest));
    }

    function step3(callFrames)
    {
        InspectorTest.addResult("Script execution paused.");
        InspectorTest.captureStackTrace(callFrames);
        InspectorTest.showScriptSource("debug-inlined-scripts-fragment-id.html", step4);
    }

    function step4(sourceFrame)
    {
        InspectorTest.dumpSourceFrameContents(sourceFrame);
        InspectorTest.resumeExecution(InspectorTest.waitUntilPaused.bind(null, step5));
    }

    function step5(callFrames)
    {
        if (callFrames[0].location.lineNumber !== 9) {
            InspectorTest.resumeExecution(InspectorTest.waitUntilPaused.bind(null, step5));
            return;
        }
        InspectorTest.addResult("Script execution paused.");
        InspectorTest.captureStackTrace(callFrames);
        InspectorTest.showScriptSource("debug-inlined-scripts-fragment-id.html", step6);
    }

    function step6(sourceFrame)
    {
        InspectorTest.dumpSourceFrameContents(sourceFrame);
        InspectorTest.resumeExecution(InspectorTest.waitUntilPaused.bind(null, step7));
    }

    function step7()
    {
        InspectorTest.resumeExecution(InspectorTest.waitUntilPaused.bind(null, step5));
    }
}

</script>

</head>

<body onload="runTest()">
<p>
Tests that all inlined scripts from the same document are shown in the same source frame with html script tags.
<a id="hash" href="https://bugs.webkit.org/show_bug.cgi?id=54544">Bug 54544.</a>
</p>

</body>
</html>
