third_party/WebKit/LayoutTests/inspector/sources/debugger/debug-inlined-scripts-fragment-id-expected.txt

Tests that all inlined scripts from the same document are shown in the same source frame with html script tags. Bug 54544.

window.location: #hash
Script source was shown.
Script execution paused.
Call stack:
    0) f4 (debug-inlined-scripts-fragment-id.html:6)
    1)  (debug-inlined-scripts-fragment-id.html:8)
==Source frame contents start==
<html>
<head>
<script>
function f4()
{
    return 0;
}
f4();
</script>

<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>

<script>

var test = function() {
  SourcesTestRunner.startDebuggerTest(step0, true);

  function step0() {
    TestRunner.evaluateInPage('window.location="#hash"', step1);
  }

  function step1(loc) {
    TestRunner.addResult('window.location: ' + loc.description);
    SourcesTestRunner.showScriptSource('debug-inlined-scripts-fragment-id.html', step2);
  }

  function step2(sourceFrame) {
    TestRunner.addResult('Script source was shown.');
    SourcesTestRunner.setBreakpoint(sourceFrame, 5, '', true);
    SourcesTestRunner.waitUntilPaused(step3);
    TestRunner.reloadPage(SourcesTestRunner.completeDebuggerTest.bind(InspectorTest));
  }

  function step3(callFrames) {
    TestRunner.addResult('Script execution paused.');
    SourcesTestRunner.captureStackTrace(callFrames);
    SourcesTestRunner.showScriptSource('debug-inlined-scripts-fragment-id.html', step4);
  }

  function step4(sourceFrame) {
    SourcesTestRunner.dumpSourceFrameContents(sourceFrame);
    SourcesTestRunner.resumeExecution(SourcesTestRunner.waitUntilPaused.bind(null, step5));
  }

  function step5(callFrames) {
    if (callFrames[0].location.lineNumber !== 9) {
      SourcesTestRunner.resumeExecution(SourcesTestRunner.waitUntilPaused.bind(null, step5));
      return;
    }
    TestRunner.addResult('Script execution paused.');
    SourcesTestRunner.captureStackTrace(callFrames);
    SourcesTestRunner.showScriptSource('debug-inlined-scripts-fragment-id.html', step6);
  }

  function step6(sourceFrame) {
    SourcesTestRunner.dumpSourceFrameContents(sourceFrame);
    SourcesTestRunner.resumeExecution(SourcesTestRunner.waitUntilPaused.bind(null, step7));
  }

  function step7() {
    SourcesTestRunner.resumeExecution(SourcesTestRunner.waitUntilPaused.bind(null, step5));
  }
};

</script>

</head>

<body onload="runTest()">
<p>
Tests that all inlined scripts from the same document are shown in the same source frame with html script tags.
<a id="hash" href="https://bugs.webkit.org/show_bug.cgi?id=54544">Bug 54544.</a>
</p>

</body>
</html>

==Source frame contents end==
Page reloaded.
